Add basic dark mode
Implementation is ugly. Styles are all over JS files and colors are repeated. But I don't how to achieve that in much cleaner / simpler way.
For now, it works.

Also, waiting for Expo SDK 37 to be able to have a different splash screen when in dark mode: https://expo.canny.io/feature-requests/p/dark-mode-splash-screen